Feature/yolov8 (#3)
* Add YOLOv8 support

* Add documentation update run.py

* Fix visualization bugs and add overlay

* Add YOLOv8

* Fix import error from refactor

* Update README and LICENSE
