+++ CHANGELOG.md
@@ -3,7 +3,18 @@
 ## Known Issues:
 - None
 
-## v0.1.1 - Nate Haddad, 1/21/2022
+## v0.3.0 - Nate Haddad, 5/5/2023
+- Add support for `YOLO-NAS` models
+- Fix text visualization alignment
+
+## v0.2.0 - Nate Haddad, 1/24/2023
+- Add support for `YOLOv8` ONNX models
+- Refactor `object_detection` module
+- Refactor `object_tracking` module
+- Add `.pylintrc`
+- Add project version
+
+## v0.1.1 - Nate Haddad, 1/21/2023
 - Fix track visualization (visualizes tracks, not only track updates)
 
 ## v0.1.0 - Nate Haddad, 12/30/2022
