+++ config/coco.yml
@@ -1,10 +1,11 @@
-data_path: data/first_test.mp4
-outfile: video_out.mp4
+data_path: data/nyc.mp4
+outfile: nyc_out.mp4
 
 
 detector:
   visualize: False
-  weights_path: models/yolov7-tiny.onnx
+  weights_path: yolo_nas_l
+  pretrained: coco
 
 
 tracker:
