+++ fast_track/object_detection/__init__.py
@@ -1,4 +1,7 @@
 """ Initializes the object_detection module """
 
 from .object_detector import ObjectDetector
-from .yolov7 import YOLOv7
+from .object_detector_onnx import ObjectDetectorONNX
+from .third_party.yolov7.yolov7 import YOLOv7
+from .third_party.yolov8.yolov8 import YOLOv8
+from .third_party.yolo_nas.yolo_nas import YOLONAS
