+++ fast_track/object_tracking/__init__.py
@@ -1,4 +1,4 @@
 """ Initializes the object_tracking module """
 
 from .object_tracker import ObjectTracker
-from .third_party.byte_tracker import BYTETracker
+from .third_party.bytetrack.byte_tracker import BYTETracker
