+++ run.py
@@ -4,7 +4,7 @@ import cv2
 import yaml
 
 from fast_track import Pipeline
-from fast_track.object_detection import YOLOv7
+from fast_track.object_detection import YOLOv8, YOLONAS
 from fast_track.object_tracking import BYTETracker
 
 
@@ -13,7 +13,7 @@ if __name__ == '__main__':
         config = yaml.safe_load(f)
 
     camera = cv2.VideoCapture(config['data_path'])
-    detector = YOLOv7(**config['detector'], names=config['names'], image_shape=(camera.get(3), camera.get(4)))
+    detector = YOLONAS(**config['detector'], names=config['names'], image_shape=(camera.get(3), camera.get(4)))
     tracker = BYTETracker(**config['tracker'], names=config['names'])
 
     with Pipeline(camera=camera, detector=detector, tracker=tracker, outfile=config['outfile']) as p:
